+++ modelslab/schemas/threed.py
@@ -27,7 +27,7 @@ class Text23D(BaseSchema):
         None,
         description="Seed for random number generation."
     )
-    guidance_scale : Optional[int] = Field(
+    guidance_scale : Optional[float] = Field(
         None,
         description="Guidance scale for the generation."
     )
